Extract the composition-aware instance validator so a second gate can reuse it
`instance_errors` and the `merged_constraints` walk beneath it were top-level
methods inside scripts/check-fixture-coverage.rb, which runs its whole check on
load — so a second gate could not require them without also running the fixture
guard. Moved verbatim into scripts/schema_instance_validator.rb, in the
scripts/bc3_route_normalizer.rb style: one definition, two callers, no parallel
validator to drift.

`module_function` plus a top-level `include` keeps every existing call site
receiverless and unchanged, so this is behaviour-preserving; the guard's own
self-test (1 positive + 8 negative + 20 synthetic cases) is what says so.

What stays in the fixture guard is what is about the fixture guard: the
concrete-instance rule and the #408 rich-text-emitter inventory.
